Split state/analytics/actions.
This helps prevent a simple bumpStat or record from pulling in a bunch
of internal and external libraries.
