+++ client/state/analytics/actions/package.json
@@ -1,0 +1,3 @@
+{
+	"sideEffects": false
+}
